remove old classes and update with our changes
