+++ inc/Dependencies/BerlinDB/Database/Base.php
@@ -22,6 +22,7 @@ defined( 'ABSPATH' ) || exit;
  *
  * @since 1.0.0
  */
+#[\AllowDynamicProperties]
 class Base {
 
 	/**
@@ -280,7 +281,7 @@ class Base {
 	 *
 	 * @since 1.0.0
 	 *
-	 * @return \wpdb Database interface, or False if not set
+	 * @return \wpdb|false Database interface, or False if not set
 	 */
 	protected function get_db() {
