+++ inc/Dependencies/BerlinDB/Database/Query.php
@@ -77,7 +77,7 @@ class Query extends Base {
 	 * @since 1.0.0
 	 * @var   string
 	 */
-	protected $table_schema = '\\WP_Rocket\Dependencies\BerlinDB\\Database\\Schema';
+	protected $table_schema = '\\WP_Rocket\\Dependencies\\BerlinDB\\Database\\Schema';
 
 	/** Item ******************************************************************/
 
@@ -113,7 +113,7 @@ class Query extends Base {
 	 * @since 1.0.0
 	 * @var   mixed
 	 */
-	protected $item_shape = '\\WP_Rocket\Dependencies\BerlinDB\\Database\\Row';
+	protected $item_shape = '\\WP_Rocket\\Dependencies\\BerlinDB\\Database\\Row';
 
 	/** Cache *****************************************************************/
 
@@ -342,12 +342,13 @@ class Query extends Base {
 	 * @since 1.0.0
 	 *
 	 * @param string|array $query Array or URL query string of parameters.
+	 * @param bool         $use_cache Use DB cache or not. (custom parameter added by us!)
 	 * @return array|int List of items, or number of items when 'count' is passed as a query var.
 	 */
-	public function query( $query = array() ) {
+	public function query( $query = array(), bool $use_cache = true ) {
 		$this->parse_query( $query );
 
-		return $this->get_items();
+		return $this->get_items( $use_cache );
 	}
 
 	/** Private Setters *******************************************************/
@@ -857,9 +858,11 @@ class Query extends Base {
 	 *
 	 * @since 1.0.0
 	 *
+	 * @param bool $use_cache Use DB cache or not. (custom parameter added by us!)
+	 *
 	 * @return array|int List of items, or number of items when 'count' is passed as a query var.
 	 */
-	private function get_items() {
+	private function get_items( bool $use_cache = true ) {
 
 		/**
 		 * Fires before object items are retrieved.
@@ -880,7 +883,7 @@ class Query extends Base {
 
 		// Check the cache
 		$cache_key   = $this->get_cache_key();
-		$cache_value = $this->cache_get( $cache_key, $this->cache_group );
+		$cache_value = $use_cache ? $this->cache_get( $cache_key, $this->cache_group ) : false;
 
 		// No cache value
 		if ( false === $cache_value ) {
@@ -889,14 +892,16 @@ class Query extends Base {
 			// Set the number of found items
 			$this->set_found_items( $item_ids );
 
-			// Format the cached value
-			$cache_value = array(
-				'item_ids'    => $item_ids,
-				'found_items' => intval( $this->found_items ),
-			);
+			if ( $use_cache ) {
+				// Format the cached value
+				$cache_value = array(
+					'item_ids'    => $item_ids,
+					'found_items' => intval( $this->found_items ),
+				);
 
-			// Add value to the cache
-			$this->cache_add( $cache_key, $cache_value, $this->cache_group );
+				// Add value to the cache
+				$this->cache_add( $cache_key, $cache_value, $this->cache_group );
+			}
 
 		// Value exists in cache
 		} else {
